add latest specs

+++ services/mgmt/billing/src/package_2017_04_preview/operations.rs
@@ -12,7 +12,7 @@ pub mod billing_periods {
         subscription_id: &str,
         filter: Option<&str>,
         skiptoken: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<BillingPeriodsListResult, list::Error> {
         let http_client = operation_config.http_client();
         let url_str = &format!(
@@ -181,7 +181,7 @@ pub mod invoices {
         expand: Option<&str>,
         filter: Option<&str>,
         skiptoken: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<InvoicesListResult, list::Error> {
         let http_client = operation_config.http_client();
         let url_str = &format!(
