+++ services/mgmt/billing/src/package_2020_05/operations.rs
@@ -6119,7 +6119,7 @@ pub mod billing_periods {
         subscription_id: &str,
         filter: Option<&str>,
         skiptoken: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
     ) -> std::result::Result<BillingPeriodsListResult, list::Error> {
         let http_client = operation_config.http_client();
         let url_str = &format!(
